+++ lcviz/helper.py
@@ -84,7 +84,13 @@ def _get_display_unit(app, axis):
         if axis == 'time':
             return u.d
         elif axis == 'flux':
-            return app._jdaviz_helper.default_time_viewer._obj.glue_viewer.data()[0].flux.unit
+            try:
+                data = app._jdaviz_helper.default_time_viewer._obj.glue_viewer.data()
+                if len(data) > 0:
+                    return data[0].flux.unit
+                return u.electron / u.s
+            except (ValueError, IndexError):
+                return u.electron / u.s
         else:
             raise ValueError(f"could not find units for axis='{axis}'")
     try:
@@ -107,7 +113,7 @@ class LCviz(ConfigHelper):
         'toolbar': ['g-data-tools', 'g-subset-tools', 'g-viewer-creator', 'g-coords-info'],
         'tray': ['g-metadata-viewer', 'flux-column',
                  'plot-options', 'g-subset-tools',
-                 'g-markers', 'time-selector',
+                 'g-markers', 'time-selector', 'photometric-extraction',
                  'stitch', 'flatten', 'frequency-analysis', 'ephemeris',
                  'binning', 'export', 'logger'],
         'viewer_area': [{'container': 'col'}]}