+++ CHANGES.rst
@@ -1,6 +1,8 @@
 1.3.0 (unreleased)
 ------------------
 
+* Photometric Extraction plugin to extract light curves from TPF cubes. [#115]
+
 1.2.2 (unreleased)
 ------------------
 
@@ -93,7 +95,6 @@
 * Default data labels no longer include flux-origin, but do include quarter/campaign/sector. [#111]
 
 * Basic stitch plugin to combine light curves into a single entry. [#107]
-
 * Metadata plugin: show undefined entries as empty string instead of object repr. [#108]
 
 * Raise error when parser can't identify file_obj [#106]